+++ CHANGELOG.rst
@@ -7,6 +7,11 @@ Note that ``PHAB_ID=#`` and ``RB_ID=#`` correspond to associated messages in com
 Unreleased
 ----------
 
+New Features
+~~~~~~~~~~~~
+
+* finagle-redis: Add `ConnectionInitCommand` stack to set database and password.
+
 Runtime Behavior Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~